+++ lib/active_interaction/filter_method.rb
@@ -7,7 +7,7 @@ module ActiveInteraction
       @method_name, @block = method_name, block
 
       @attribute = args.shift if args.first.is_a?(Symbol)
-      @options   = (args.first || {}).dup
+      @options = (args.first || {}).dup
 
       if @options.include?(:default)
         raise ArgumentError, ':default is not supported inside filter blocks'