+++ spec/active_interaction/integration/model_interaction_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
 
-describe 'ModelInteraciton' do
+describe 'ModelInteraction' do
   it_behaves_like 'an interaction', :model, -> { Proc.new {} }, class: Proc
 end