+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.1.2
 
-- `execute` will now have the filtered version of the values passed to `run` or `run!` as was intended.
+- `execute` will now have the filtered version of the values passed
+  to `run` or `run!` as was intended.
 
 # 0.1.1